feat: remove `LogStorageAppender` and write directly to log storage
